backend/inventory_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.5.1
pymongo==4.6.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.3
